+++ lib/ModelMixins/TableMixin.ts
@@ -1,4 +1,4 @@
-import { action, computed, observable, runInAction } from "mobx";
+import { action, computed, observable, runInAction, toJS } from "mobx";
 import { createTransformer } from "mobx-utils";
 import DeveloperError from "terriajs-cesium/Source/Core/DeveloperError";
 import JulianDate from "terriajs-cesium/Source/Core/JulianDate";
@@ -10,7 +10,6 @@ import Entity from "terriajs-cesium/Source/DataSources/Entity";
 import ImageryLayerFeatureInfo from "terriajs-cesium/Source/Scene/ImageryLayerFeatureInfo";
 import { ChartPoint } from "../Charts/ChartData";
 import getChartColorForId from "../Charts/getChartColorForId";
-import AsyncLoader from "../Core/AsyncLoader";
 import Constructor from "../Core/Constructor";
 import filterOutUndefined from "../Core/filterOutUndefined";
 import isDefined from "../Core/isDefined";
@@ -21,8 +20,6 @@ import TerriaError from "../Core/TerriaError";
 import MapboxVectorTileImageryProvider from "../Map/MapboxVectorTileImageryProvider";
 import RegionProvider from "../Map/RegionProvider";
 import JSRegionProviderList from "../Map/RegionProviderList";
-import { ImageryParts } from "./MappableMixin";
-import { calculateDomain, ChartAxis, ChartItem } from "../Models/Chartable";
 import CommonStrata from "../Models/CommonStrata";
 import Model from "../Models/Model";
 import SelectableDimensions, {
@@ -34,12 +31,16 @@ import TableColumn from "../Table/TableColumn";
 import TableColumnType from "../Table/TableColumnType";
 import TableStyle from "../Table/TableStyle";
 import TableTraits from "../Traits/TableTraits";
-import MappableMixin from "./MappableMixin";
+import ChartableMixin, {
+  calculateDomain,
+  ChartAxis,
+  ChartItem
+} from "./ChartableMixin";
 import DiscretelyTimeVaryingMixin, {
   DiscreteTimeAsJS
 } from "./DiscretelyTimeVaryingMixin";
 import ExportableMixin, { ExportData } from "./ExportableMixin";
-import TimeVarying from "./TimeVarying";
+import { ImageryParts } from "./MappableMixin";
 
 // TypeScript 3.6.3 can't tell JSRegionProviderList is a class and reports
 //   Cannot use namespace 'JSRegionProviderList' as a type.ts(2709)
@@ -47,17 +48,15 @@ import TimeVarying from "./TimeVarying";
 class RegionProviderList extends JSRegionProviderList {}
 function TableMixin<T extends Constructor<Model<TableTraits>>>(Base: T) {
   abstract class TableMixin
-    extends ExportableMixin(MappableMixin(DiscretelyTimeVaryingMixin(Base)))
-    implements SelectableDimensions, TimeVarying {
+    extends ExportableMixin(ChartableMixin(DiscretelyTimeVaryingMixin(Base)))
+    implements SelectableDimensions {
     get hasTableMixin() {
       return true;
     }
-    /**
-     * The raw data table in column-major format, i.e. the outer array is an
-     * array of columns.
-     */
+
+    // Always use the getter and setter for this
     @observable
-    dataColumnMajor: string[][] | undefined;
+    protected _dataColumnMajor: string[][] | undefined;
 
     /**
      * The list of region providers to be used with this table.
@@ -65,7 +64,43 @@ function TableMixin<T extends Constructor<Model<TableTraits>>>(Base: T) {
     @observable
     regionProviderList: RegionProviderList | undefined;
 
-    private _dataLoader = new AsyncLoader(this.forceLoadTableMixin.bind(this));
+    /**
+     * The raw data table in column-major format, i.e. the outer array is an
+     * array of columns.
+     */
+    @computed
+    get dataColumnMajor(): string[][] | undefined {
+      const dataColumnMajor = this._dataColumnMajor;
+      if (
+        this.removeDuplicateRows &&
+        dataColumnMajor !== undefined &&
+        dataColumnMajor.length >= 1
+      ) {
+        // De-duplication is slow and memory expensive, so should be avoided if possible.
+        const rowsToRemove = new Set();
+        const seenRows = new Set();
+        for (let i = 0; i < dataColumnMajor[0].length; i++) {
+          const row = dataColumnMajor.map(col => col[i]).join();
+          if (seenRows.has(row)) {
+            // Mark row for deletion
+            rowsToRemove.add(i);
+          } else {
+            seenRows.add(row);
+          }
+        }
+
+        if (rowsToRemove.size > 0) {
+          return dataColumnMajor.map(col =>
+            col.filter((cell, idx) => !rowsToRemove.has(idx))
+          );
+        }
+      }
+      return dataColumnMajor;
+    }
+
+    set dataColumnMajor(newDataColumnMajor: string[][] | undefined) {
+      this._dataColumnMajor = newDataColumnMajor;
+    }
 
     /**
      * Gets a {@link TableColumn} for each of the columns in the raw data.
@@ -467,9 +502,27 @@ function TableMixin<T extends Constructor<Model<TableTraits>>>(Base: T) {
       return this.tableColumns.find(column => column.name === name);
     }
 
-    protected abstract forceLoadTableData(): Promise<string[][]>;
+    protected async forceLoadMapItems() {
+      const dataColumnMajor = await this.forceLoadTableData();
 
-    protected async loadRegionProviderList() {
+      if (dataColumnMajor !== undefined && dataColumnMajor !== null) {
+        runInAction(() => {
+          this.dataColumnMajor = dataColumnMajor;
+        });
+      }
+    }
+
+    /**
+     * Forces load of the table data. This method does _not_ need to consider
+     * whether the table data is already loaded.
+     *
+     * It is guaranteed that `loadMetadata` has finished before this is called, and `regionProviderList` is set.
+     *
+     * You **can not** make changes to observables until **after** an asynchronous call {@see AsyncLoader}.
+     */
+    protected abstract forceLoadTableData(): Promise<string[][] | undefined>;
+
+    async loadRegionProviderList() {
       if (isDefined(this.regionProviderList)) return;
 
       const regionProvidersPromise:
@@ -483,31 +536,9 @@ function TableMixin<T extends Constructor<Model<TableTraits>>>(Base: T) {
       runInAction(() => (this.regionProviderList = regionProvidersPromise));
     }
 
-    private async forceLoadTableMixin(): Promise<void> {
-      await this.loadRegionProviderList();
-
-      const dataColumnMajor = await this.forceLoadTableData();
-      runInAction(() => {
-        this.dataColumnMajor = dataColumnMajor;
-      });
-    }
-
-    protected forceLoadChartItems(force?: boolean) {
-      return this._dataLoader.load(force);
-    }
-
-    protected forceLoadMapItems(force?: boolean) {
-      return this._dataLoader.load(force);
-    }
-
-    dispose() {
-      super.dispose();
-      this._dataLoader.dispose();
-    }
-
     /*
      * Appends new table data in column major format to this table.
-     * It is assumed that thhe column order is the same for both the tables.
+     * It is assumed that the column order is the same for both the tables.
      */
     @action
     append(dataColumnMajor2: string[][]) {
@@ -722,7 +753,10 @@ function TableMixin<T extends Constructor<Model<TableTraits>>>(Base: T) {
               return undefined;
             }
           }),
-          show: this.show
+          show: this.show,
+          clippingRectangle: this.clipToRectangle
+            ? this.cesiumRectangle
+            : undefined
         };
       }
     );