+++ test/Models/WebMapServiceCatalogItemSpec.ts
@@ -1,4 +1,6 @@
 import { autorun, runInAction } from "mobx";
+import GeographicTilingScheme from "terriajs-cesium/Source/Core/GeographicTilingScheme";
+import WebMercatorTilingScheme from "terriajs-cesium/Source/Core/WebMercatorTilingScheme";
 import WebMapServiceImageryProvider from "terriajs-cesium/Source/Scene/WebMapServiceImageryProvider";
 import { ImageryParts } from "../../lib/ModelMixins/MappableMixin";
 import CommonStrata from "../../lib/Models/CommonStrata";
@@ -31,6 +33,34 @@ describe("WebMapServiceCatalogItem", function() {
       );
     });
     return wms.loadMapItems();
+  });
+
+  describe("selects correct tilingScheme", () => {
+    it("uses 4326 is no 3857", async function() {
+      const terria = new Terria();
+      const wms = new WebMapServiceCatalogItem("test", terria);
+      runInAction(() => {
+        wms.setTrait("definition", "url", "test/WMS/wms_crs.xml");
+        wms.setTrait("definition", "layers", "ls8_nbart_geomedian_annual");
+      });
+      await wms.loadMapItems();
+
+      expect(wms.crs).toBe("EPSG:4326");
+      expect(wms.tilingScheme instanceof GeographicTilingScheme).toBeTruthy();
+    });
+
+    it("uses 3857 over 4326", async function() {
+      const terria = new Terria();
+      const wms = new WebMapServiceCatalogItem("test", terria);
+      runInAction(() => {
+        wms.setTrait("definition", "url", "test/WMS/wms_nested_groups.xml");
+        wms.setTrait("definition", "layers", "ls8_nbart_geomedian_annual");
+      });
+      await wms.loadMapItems();
+
+      expect(wms.crs).toBe("EPSG:3857");
+      expect(wms.tilingScheme instanceof WebMercatorTilingScheme).toBeTruthy();
+    });
   });
 
   it("updates description from a GetCapabilities", async function() {